App server

+++ tsrc/example1/src/docker.run.sh
@@ -1,16 +1,22 @@
 #!/bin/bash
 
 CONTAINER_NAME=$1
+FOLDER=$2
 
 if [ -z $CONTAINER_NAME ]; then
     echo "Where is CONTAINER_NAME dude? Exiting" 1>&2
     exit
 fi
 
+if [ -z $FOLDER ]; then
+    echo "Where is FOLDER dude? Exiting" 1>&2
+    exit
+fi
+
 docker stop $CONTAINER_NAME
 docker rm -f $CONTAINER_NAME
 
-./docker.build.sh $CONTAINER_NAME
+./docker.build.sh $CONTAINER_NAME $FOLDER
 
 KEYS="-it"
